Move TreeTableRecordItem class into separate file
